Testing new constructor
